Allow control of MOUNTPOINT and of root for operations
If $MOUNTPOINT is set, use that value instead of calculating it based on
$DATASET.  This allows the test script to be used for testing
filesystems with different mountpoint conventions than ZFS.

If $ROOTDIR is set, perform operations there instead of $MOUNTPOINT.
This allows files and directories created by the testing to be
segregated, both for easy cleanup and to prevent files on the file
system that might not be disposable, from being removed by the test.

Previously, logs were always placed in the same directory as
runstress.sh (and other) scripts.  This change allows the user to set
LOGDIR, in which case the log files will be placed there.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Closes #5