Use default for randsleep
Most functions loop with a sleep of random duration (up to some maximum)
once per iteration.

The randsleep function has a default maximum of 60 seconds, but the
functions were specifying a maximum sleep time of 60 seconds anyway.

This commit changes those functions not to specify a maximum and use the
default, so the sleep time of all functions can be adjusted by altering
a single value.