Simplify for loop iteration termination conditions
Several tests use a for() loop to perform some action a random number of
times.

The termination condition of the loops contained a reference to $RANDOM.
Because this expression is evaluated after each pass through the loop,
the index was compared against a different random number each time.  The
actual behavior was thus more like a while() loop with a random
termination condition than a typical for loop.

This patch evalutes $RANDOM once to determine the number of passes
through the loop, so the code's behavior is more obvious.